# RoundRobinLoadBalancer to allow duplicate addresses