--- a/servicetalk-benchmarks/build.gradle
+++ b/servicetalk-benchmarks/build.gradle
@@ -29,6 +29,7 @@ dependencies {
   implementation project(":servicetalk-http-api")
   implementation project(":servicetalk-http-netty")
   implementation project(":servicetalk-transport-netty-internal")
+  implementation project(":servicetalk-loadbalancer")
   implementation "com.google.code.findbugs:jsr305:$jsr305Version"
   implementation "io.netty:netty-codec-http:$nettyVersion"
   implementation "org.openjdk.jmh:jmh-core:$jmhCoreVersion"