--- a/servicetalk-loadbalancer/src/test/java/io/servicetalk/loadbalancer/RoundRobinLoadBalancerTest.java
+++ b/servicetalk-loadbalancer/src/test/java/io/servicetalk/loadbalancer/RoundRobinLoadBalancerTest.java
@@ -243,6 +243,15 @@ public class RoundRobinLoadBalancerTest {
     }
 
     @Test
+    public void addressIsAddedTwice() {
+        assertThat(lb.activeAddresses(), is(empty()));
+        sendServiceDiscoveryEvents(upEvent("address-1"));
+        assertThat(lb.activeAddresses(), hasSize(1));
+        sendServiceDiscoveryEvents(upEvent("address-1"));
+        assertThat(lb.activeAddresses(), hasSize(2));
+    }
+
+    @Test
     public void noServiceDiscoveryEvent() {
         selectConnectionListener.listen(lb.selectConnection(any()));
         selectConnectionListener.verifyFailure(NoAvailableHostException.class);