Motivation:
RoundRobinLoadBalancer currently prevents duplicates from being
inserted. However the ServiceDiscoverer is responsible for filtering
duplicates in the typical use case. Allowing duplicates at the LB layer
allows to give more weight to specific addresses if desired.

Modifications:
- RoundRobinLoadBalancer simplify addition/removal to forgo duplicate
detection

Result:
RoundRobinLoadBalancer code simplified to ignore duplicate detection,
addition doesn't require linear search.