@Scottmitch 2020-05-14
```
## this pr
Benchmark                                          (ops)   Mode  Cnt        Score       Error  Units
RoundRobinLoadBalancerSDEventsBenchmark.available      5  thrpt    5  2856870.483 ± 17853.380  ops/s
RoundRobinLoadBalancerSDEventsBenchmark.available     10  thrpt    5  1644923.209 ± 26508.869  ops/s
RoundRobinLoadBalancerSDEventsBenchmark.available    100  thrpt    5   143425.550 ± 10880.633  ops/s
RoundRobinLoadBalancerSDEventsBenchmark.mixed          5  thrpt    5  2151777.842 ± 51603.588  ops/s
RoundRobinLoadBalancerSDEventsBenchmark.mixed         10  thrpt    5  1249466.523 ± 27359.906  ops/s
RoundRobinLoadBalancerSDEventsBenchmark.mixed        100  thrpt    5    55072.715 ±   530.522  ops/s 

## master
Benchmark                                          (ops)   Mode  Cnt        Score       Error  Units
RoundRobinLoadBalancerSDEventsBenchmark.available      5  thrpt    5  1809302.630 ± 47720.945  ops/s
RoundRobinLoadBalancerSDEventsBenchmark.available     10  thrpt    5   768544.812 ± 23195.555  ops/s
RoundRobinLoadBalancerSDEventsBenchmark.available    100  thrpt    5    11612.291 ±   262.869  ops/s
RoundRobinLoadBalancerSDEventsBenchmark.mixed          5  thrpt    5  1911995.403 ± 35511.624  ops/s
RoundRobinLoadBalancerSDEventsBenchmark.mixed         10  thrpt    5   810828.434 ± 10277.926  ops/s
RoundRobinLoadBalancerSDEventsBenchmark.mixed        100  thrpt    5    17890.800 ±   452.113  ops/s
```